JAVA-590: fixed test to work with older server versions.  Using Collections.addAll

+++ src/test/com/mongodb/JavaClientTest.java
@@ -564,6 +564,10 @@ public class JavaClientTest extends TestCase {
     
     @Test
     public void testAggregation(){
+        if (!serverIsAtLeastVersion(2.1)) {
+            return;
+        }
+
         DBCollection c = _db.getCollection( "aggregationTest" );
         c.drop();
